requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
mysql-connector-python==8.0.27
protobuf==3.19.1
pycodestyle==2.8.0
pyflakes==2.4.0
ratelimit==2.2.1
requests==2.27.1
selenium==3.141.0
tqdm==4.62.3
urllib3==1.26.7
